requirements/local.txt
-r ./base.txt

Werkzeug==0.14.1  # https://github.com/pallets/werkzeug
ipdb==0.11  # https://github.com/gotcha/ipdb
Sphinx==1.7.4  # https://github.com/sphinx-doc/sphinx
psycopg2==2.7.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
pytest==3.5.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.1  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8==3.5.0  # https://github.com/PyCQA/flake8
coverage==4.5.1  # https://github.com/nedbat/coveragepy
pylint==1.9.0
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.11
pylint-plugin-utils==0.2.6

# Django
# ------------------------------------------------------------------------------
factory-boy==2.11.1  # https://github.com/FactoryBoy/factory_boy
django-test-plus==1.0.22  # https://github.com/revsys/django-test-plus

django-debug-toolbar==1.9.1  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.0.7  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.5.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.2.1  # https://github.com/pytest-dev/pytest-django
